gpu_PYTHON_distribution/requirement.txt
cupy-cuda12x 
numpy
gdal
matplotlib